Cleaned services, reduced lines of code, implementation of validation (incomplete)

+++ src/main/java/com/jjo/h2/repositories/TagsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.jjo.h2.model.Tags;
 
-public interface TagsRepository extends JpaRepository<Tags, Integer> {
+public interface TagsRepository extends JpaRepository<Tags, Long> {
 
 }
